move correlation plot test to test area
